Move LimitClause to pkg/ast
